+++ src/pulumi/core-dtap/analytics/databricks/workspaces/engineering.py
@@ -1,26 +1,25 @@
-import os
+from os import getenv
+
 import pulumi
 import pulumi_azure_native as azure_native
 import pulumi_azuread as azuread
-
-from pulumi_databricks import Provider as DatabricksProvider
-from pulumi_databricks import ProviderArgs as DatabricksProviderArgs
-from pulumi_databricks import databricks
-
 from ingenii_azure_data_platform.iam import (
     GroupRoleAssignment,
     ServicePrincipalRoleAssignment,
 )
 from ingenii_azure_data_platform.logs import log_diagnostic_settings
 from ingenii_azure_data_platform.utils import generate_hash, generate_resource_name
-
-from project_config import azure_client, platform_config, platform_outputs, DTAP_ROOT
-from platform_shared import add_config_registry_secret
 from logs import log_analytics_workspace
 from management import resource_groups
 from management.user_groups import user_groups
-from security import credentials_store
 from network import vnet
+from platform_shared import add_config_registry_secret
+from project_config import DTAP_ROOT, azure_client, platform_config, platform_outputs
+from pulumi import FileAsset, ResourceOptions
+from pulumi_databricks import Provider as DatabricksProvider
+from pulumi_databricks import ProviderArgs as DatabricksProviderArgs
+from pulumi_databricks import databricks
+from security import credentials_store
 from storage.datalake import datalake, datalake_containers
 
 outputs = platform_outputs["analytics"]["databricks"]["workspaces"]["engineering"] = {}
@@ -67,6 +66,9 @@ workspace = azure_native.databricks.Workspace(
     ),
     sku=azure_native.databricks.SkuArgs(name="Premium"),
     resource_group_name=resource_groups["infra"].name,
+    opts=ResourceOptions(
+        protect=platform_config.resource_protection,
+    ),
 )
 
 outputs["name"] = workspace.name
@@ -109,10 +111,12 @@ for assignment in workspace_config.get("iam", {}).get("role_assignments", []):
 databricks_provider = DatabricksProvider(
     resource_name=workspace_name,
     args=DatabricksProviderArgs(
-        azure_client_id=os.getenv("ARM_CLIENT_ID", azure_client.client_id),
-        azure_client_secret=os.getenv("ARM_CLIENT_SECRET"),
-        azure_subscription_id=os.getenv("ARM_SUBSCRIPTION_ID", azure_client.subscription_id),
-        azure_tenant_id=os.getenv("ARM_TENANT_ID", azure_client.tenant_id),
+        azure_client_id=getenv("ARM_CLIENT_ID", azure_client.client_id),
+        azure_client_secret=getenv("ARM_CLIENT_SECRET"),
+        azure_subscription_id=getenv(
+            "ARM_SUBSCRIPTION_ID", azure_client.subscription_id
+        ),
+        azure_tenant_id=getenv("ARM_TENANT_ID", azure_client.tenant_id),
         azure_workspace_resource_id=workspace.id,
     ),
 )
@@ -123,10 +127,14 @@ databricks_provider = DatabricksProvider(
 databricks.WorkspaceConf(
     resource_name=workspace_name,
     custom_config={
-        "enableDcs": workspace_config["config"].get("enable_container_services", "false"),
-        "enableIpAccessLists": workspace_config["config"].get("enable_ip_access_lists", "false"),
+        "enableDcs": workspace_config["config"].get(
+            "enable_container_services", "false"
+        ),
+        "enableIpAccessLists": workspace_config["config"].get(
+            "enable_ip_access_lists", "false"
+        ),
     },
-    opts=pulumi.ResourceOptions(provider=databricks_provider),
+    opts=ResourceOptions(provider=databricks_provider),
 )
 # ----------------------------------------------------------------------------------------------------------------------
 # ENGINEERING DATABRICKS WORKSPACE -> SECRETS & TOKENS
@@ -137,7 +145,7 @@ secret_scope_name = "main"
 secret_scope = databricks.SecretScope(
     resource_name=f"{workspace_name}-secret-scope-main",
     name=secret_scope_name,
-    opts=pulumi.ResourceOptions(provider=databricks_provider),
+    opts=ResourceOptions(provider=databricks_provider),
 )
 
 # DBT TOKEN
@@ -148,7 +156,7 @@ dbt_token_resource_name = f"{workspace_name}-token-for-dbt"
 dbt_token = databricks.Token(
     resource_name=dbt_token_resource_name,
     comment="Data Build Tool Token - Used for DBT automation",
-    opts=pulumi.ResourceOptions(provider=databricks_provider),
+    opts=ResourceOptions(provider=databricks_provider),
 )
 
 dbt_token_as_scope_secret = databricks.Secret(
@@ -156,7 +164,7 @@ dbt_token_as_scope_secret = databricks.Secret(
     scope=secret_scope.id,
     string_value=dbt_token.token_value,
     key=dbt_token_name,
-    opts=pulumi.ResourceOptions(provider=databricks_provider),
+    opts=ResourceOptions(provider=databricks_provider),
 )
 
 dbt_token_as_key_vault_secret = azure_native.keyvault.Secret(
@@ -173,7 +181,7 @@ dbt_token_as_key_vault_secret = azure_native.keyvault.Secret(
 datafactory_token = databricks.Token(
     resource_name=f"{workspace_name}-token-for-datafactory",
     comment="Data Factory Token - Used for Data Factory integration",
-    opts=pulumi.ResourceOptions(provider=databricks_provider),
+    opts=ResourceOptions(provider=databricks_provider),
 )
 
 # ----------------------------------------------------------------------------------------------------------------------
@@ -243,7 +251,7 @@ storage_mounts_dbw_password = databricks.Secret(
     scope=secret_scope.id,
     string_value=storage_mounts_sp_password.value,
     key=storage_mounts_sp_name,
-    opts=pulumi.ResourceOptions(provider=databricks_provider),
+    opts=ResourceOptions(provider=databricks_provider),
 )
 
 # ----------------------------------------------------------------------------------------------------------------------
@@ -273,7 +281,7 @@ for definition in workspace_config.get("storage_mounts", []):
         container_name=definition["container_name"],
         mount_name=definition["mount_name"],
         cluster_id=system_cluster.id,
-        opts=pulumi.ResourceOptions(
+        opts=ResourceOptions(
             provider=databricks_provider,
             delete_before_replace=True,
         ),
@@ -284,7 +292,7 @@ for definition in workspace_config.get("storage_mounts", []):
 # ----------------------------------------------------------------------------------------------------------------------
 
 blob_name = "pre_process-1.0.0-py3-none-any.whl"
-pre_processing_package = pulumi.FileAsset(f"{DTAP_ROOT}/assets/{blob_name}")
+pre_processing_package = FileAsset(f"{DTAP_ROOT}/assets/{blob_name}")
 pre_processing_blob = azure_native.storage.Blob(
     resource_name=f"{workspace_name}-pre_processing_package",
     account_name=datalake.name,
@@ -292,7 +300,7 @@ pre_processing_blob = azure_native.storage.Blob(
     container_name=datalake_containers["utilities"].name,
     resource_group_name=resource_groups["data"].name,
     source=pre_processing_package,
-    opts=pulumi.ResourceOptions(
+    opts=ResourceOptions(
         ignore_changes=["content_md5"], depends_on=[storage_mounts["utilities"]]
     ),
 )
@@ -326,13 +334,14 @@ for ref_key, cluster_config in workspace_config.get("clusters", {}).items():
         for lib in libraries.get("pypi", [])
     ]
     whl_libraries_str = libraries.get("whl", [])
-    
+
     if ref_key == "default":
-        whl_libraries_str.append("dbfs:/mnt/utilities/pre_process/pre_process-1.0.0-py3-none-any.whl")
-    
+        whl_libraries_str.append(
+            "dbfs:/mnt/utilities/pre_process/pre_process-1.0.0-py3-none-any.whl"
+        )
+
     whl_libraries = [
-        databricks.ClusterLibraryArgs(whl=lib)
-        for lib in set(whl_libraries_str)
+        databricks.ClusterLibraryArgs(whl=lib) for lib in set(whl_libraries_str)
     ]
 
     cluster_libraries = pypi_libraries + whl_libraries
@@ -417,7 +426,7 @@ for ref_key, cluster_config in clusters.items():
                 permission_level="CAN_ATTACH_TO", group_name="users"
             )
         ],
-        opts=pulumi.ResourceOptions(provider=databricks_provider),
+        opts=ResourceOptions(provider=databricks_provider),
     )
 
 # ----------------------------------------------------------------------------------------------------------------------
@@ -425,6 +434,8 @@ for ref_key, cluster_config in clusters.items():
 # ----------------------------------------------------------------------------------------------------------------------
 
 add_config_registry_secret(
-    "databricks-engineering-workspace-hostname", workspace.workspace_url)
+    "databricks-engineering-workspace-hostname", workspace.workspace_url
+)
 add_config_registry_secret(
-    "databricks-engineering-cluster-name", clusters["default"].cluster_name)
+    "databricks-engineering-cluster-name", clusters["default"].cluster_name
+)