+++ src/pulumi/core-shared/management/resource_groups.py
@@ -1,6 +1,6 @@
-from project_config import platform_config, platform_outputs
-from ingenii_azure_data_platform.management import ResourceGroup
 from ingenii_azure_data_platform.iam import GroupRoleAssignment
+from ingenii_azure_data_platform.management import ResourceGroup
+from project_config import platform_config, platform_outputs
 
 from .user_groups import user_groups